requirements.txt
python-dotenv
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pyautogen
pymupdf
tiktoken
streamlit
langchain-text-splitters
